Break up the setting in the constructor
I would like to suggest breaking up the setters in the constructor. Instead of having it all checked, why not just always set the values, defaulting to null. There null either way. This way it saves time using the constructor. If for example i don't set the `$to` because i want to do it later on trough looping a list. I have to do all the setters manually. With this pull request i would only have to set the `Personalization` manually.