UTF8 force
Proposing a force to UTF8 encoding on set for content value and message subject.  If you pass non-UTF8 into the v3 API currently it will error with no reason giving; a "null" response with zero feedback to the caller.  My "experience" is I had customer data with awkward encoding which caused a rabbit hole chase to find the core issue.  I believe there is a good solid argument that my app should have forced the encoding.  I agree with this.  However, this fix at the SDK level may save someone else the pain.  Also interested in other ways to work around it and open to feedback.