Refactor Code
I believe that most of the functions I removed were redundant. Causing an overly complicated code. Not only for writing this API but for using it. Most people (probably all) will write PHP code connected to some sort of RDBMS. Which means they will pull back objects or an array. In this case I am using array as arrays are preferred in most API I have seen. There are a lot of "new Collection" functions that shouldn't exist. Also there was a lot of random syntax errors and things that don't make sense in regards to posting ("important-key", "important value) but then in the collection it was (["important key" => "important value"]) I think it was just overlooked. However I believe this is the best way to go. This reduces programming time to make the API as well as a friendly environment for integration.