Fix #644: Remove loader.php
loader.php appears to be unused, except for a stray reference in
CONTRIBUTING.md. Actually, that reference is misleading, since composer
is required in order to provide the dependency on
sendgrid/php-http-client. Because composer is required to provide the
dependency, it doesn't make sense to maintain a loader separate from the
one composer generates for us. Therefore, we should remove loader.php.